[TIMOB-25752] Fix Ti.Media constants
