+++ iphone/Classes/TiMediaAudioRecorderProxy.h
@@ -7,14 +7,9 @@
 #ifdef USE_TI_MEDIAAUDIORECORDER
 
 #import "TiFile.h"
+#import "TiMediaTypes.h"
 #import "TiProxy.h"
 #import <AVFoundation/AVAudioRecorder.h>
-
-typedef enum {
-  RecordStarted = 0,
-  RecordStopped = 1,
-  RecordPaused = 2
-} RecorderState;
 
 @interface TiMediaAudioRecorderProxy : TiProxy <AVAudioRecorderDelegate> {
   @private
